Share the latest site so that it can be referred easily
We need to have an easy way to access the current site's objects from
filters and custom helpers. It isn't sensible to pass the site to
filters as an argument: the injection chain would be deep (Site -> Item
-> ItemRep -> Filter).

In addition, the shared reference to site will be used by filters to
determine the configured temporary directory.